# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/books": {
      "post": {
        "operationId": "BooksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new book",
        "tags": [
          "books"
        ]
      },
      "get": {
        "operationId": "BooksController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all books",
        "tags": [
          "books"
        ]
      }
    },
    "/books/{id}": {
      "get": {
        "operationId": "BooksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a specific book",
        "tags": [
          "books"
        ]
      },
      "patch": {
        "operationId": "BooksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a specific book",
        "tags": [
          "books"
        ]
      },
      "delete": {
        "operationId": "BooksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a book",
        "tags": [
          "books"
        ]
      }
    }
  },
  "info": {
    "title": "Books API",
    "description": "The crude book API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the book",
            "example": "1984"
          },
          "author": {
            "type": "string",
            "description": "The author of the book",
            "example": "George Orwell"
          },
          "publicationYear": {
            "type": "number",
            "description": "The year the book was published",
            "example": 1949
          }
        },
        "required": [
          "title",
          "author",
          "publicationYear"
        ]
      },
      "UpdateBookDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the book",
            "example": "1984"
          },
          "author": {
            "type": "string",
            "description": "The author of the book",
            "example": "George Orwell"
          },
          "publicationYear": {
            "type": "number",
            "description": "The year the book was published",
            "example": 1949
          }
        }
      }
    }
  }
}